Added headers for Attack and Interface
Made changes to Interface.cpp mainly changing method names.
